+++ Modules/Microsoft365DSC/DSCResources/MSFT_IntuneAppProtectionPolicyAndroid/MSFT_IntuneAppProtectionPolicyAndroid.psm1
@@ -46,30 +46,6 @@ function Get-TargetResource
 
         [Parameter()]
         [System.Boolean]
-        $IsAssigned,
-
-        [Parameter()]
-        [System.String]
-        $ManagedBrowser,
-
-        [Parameter()]
-        [System.String]
-        $MinimumRequiredAppVersion,
-
-        [Parameter()]
-        [System.String]
-        $MinimumWarningAppVersion,
-
-        [Parameter()]
-        [System.String]
-        $MinimumRequiredOSVersion,
-
-        [Parameter()]
-        [System.String]
-        $MinimumWarningOSVersion,
-
-        [Parameter()]
-        [System.Boolean]
         $ManagedBrowserToOpenLinksRequired,
 
         [Parameter()]
@@ -125,6 +101,11 @@ function Get-TargetResource
         $FingerprintBlocked,
 
         [Parameter()]
+        [ValidateSet('allApps', 'allMicrosoftApps', 'allCoreMicrosoftApps', 'selectedPublicApps' )]
+        [System.String]
+        $AppGroupType,
+
+        [Parameter()]
         [System.String[]]
         $Apps,
 
@@ -162,7 +143,8 @@ function Get-TargetResource
         $CertificateThumbprint
     )
     Write-Verbose -Message "Checking for the Intune Android App Protection Policy {$DisplayName}"
-    $ConnectionMode = New-M365DSCConnection -Workload 'MicrosoftGraph' `
+    # beta profile as previous graph call used beta and it returns more data
+    $ConnectionMode = New-M365DSCConnection -Workload 'MicrosoftGraph' -ProfileName beta `
         -InboundParameters $PSBoundParameters
 
     #Ensure the proper dependencies are installed in the current environment.
@@ -170,9 +152,7 @@ function Get-TargetResource
 
     #region Telemetry
     $ResourceName = $MyInvocation.MyCommand.ModuleName -replace "MSFT_", ""
-    write-host 'resourcename:' $ResourceName
     $CommandName  = $MyInvocation.MyCommand
-    write-host 'commandname:' $commandname
     $data = Format-M365DSCTelemetryParameters -ResourceName $ResourceName `
         -CommandName $CommandName `
         -Parameters $PSBoundParameters
@@ -201,84 +181,75 @@ function Get-TargetResource
             throw "Multiple Policies with same displayname identified - Module currently only functions with unique names"
         }
 
-        # It may be possible to remove this once basic functionality is in place - output types differ to the graph module object though so implementing basic functionality first
-        $policy = Get-M365DSCIntuneAppProtectionPolicyAndroid -PolicyId $policyInfo.Id
         Write-Verbose -Message "Found Android App Protection Policy {$DisplayName}"
 
         $appsArray = @()
-        if ($null -ne $policy.Apps)
+        if ($null -ne $policyInfo.Apps)
         {
-            foreach ($app in $policy.Apps)
+            foreach ($app in $policyInfo.Apps)
             {
-                $appsArray += $app.mobileAppIdentifier.packageId
+                $appsArray += $app.MobileAppIdentifier.AdditionalProperties.packageId
             }
         }
 
         $assignmentsArray = @()
-        if ($null -ne $policy.Assignments)
-        {
-            $allAssignments = $policy.Assignments.target | Where-Object -FilterScript { $_.'@odata.type' -eq '#microsoft.graph.groupAssignmentTarget' }
-
-            foreach ($assignment in $allAssignments)
-            {
-                $assignmentsArray += $assignment.groupId
-            }
-        }
-
         $exclusionArray = @()
-        if ($null -ne $policy.Assignments)
+        if ($null -ne $policyInfo.Assignments)
         {
-            $allExclusions = $policy.Assignments.target | Where-Object -FilterScript { $_.'@odata.type' -eq '#microsoft.graph.exclusionGroupAssignmentTarget' }
-
-            foreach ($exclusion in $allExclusions)
+            foreach ($assignment in $policyInfo.Assignments)
             {
-                $exclusionArray += $exclusion.groupId
+                switch ($assignment.Target.AdditionalProperties.'@odata.type')
+                {
+                    '#microsoft.graph.groupAssignmentTarget'
+                    {
+                        $assignmentsArray += $assignment.Target.AdditionalProperties.groupId
+                    }
+
+                    '#microsoft.graph.exclusionGroupAssignmentTarget'
+                    {
+                        $exclusionArray += $assignment.Target.AdditionalProperties.groupId
+                    }
+                }
             }
         }
-        return @{
-            DisplayName                             = $policyInfo.DisplayName
-            Description                             = $policy.Description
-            PeriodOfflineBeforeAccessCheck          = $policy.PeriodOfflineBeforeAccessCheck
-            PeriodOnlineBeforeAccessCheck           = $policy.PeriodOnlineBeforeAccessCheck
-            AllowedInboundDataTransferSources       = $policy.AllowedInboundDataTransferSources
-            AllowedOutboundDataTransferDestinations = $policy.AllowedOutboundDataTransferDestinations
-            OrganizationalCredentialsRequired       = $policy.OrganizationalCredentialsRequired
-            AllowedOutboundClipboardSharingLevel    = $policy.AllowedOutboundClipboardSharingLevel
-            DataBackupBlocked                       = $policy.DataBackupBlocked
-            DeviceComplianceRequired                = $policy.DeviceComplianceRequired
-            IsAssigned                              = $policy.IsAssigned
-            ManagedBrowser                          = $policy.ManagedBrowser
-            MinimumRequiredAppVersion               = $policy.MinimumRequiredAppVersion
-            MinimumRequiredOSVersion                = $policy.MinimumRequiredOSVersion
-            MinimumWarningAppVersion                = $policy.MinimumWarningAppVersion
-            MinimumWarningOSVersion                 = $policy.MinimumWarningOSVersion
-            ManagedBrowserToOpenLinksRequired       = $policy.ManagedBrowserToOpenLinksRequired
-            SaveAsBlocked                           = $policy.SaveAsBlocked
-            PeriodOfflineBeforeWipeIsEnforced       = $policy.PeriodOfflineBeforeWipeIsEnforced
-            PinRequired                             = $policy.PinRequired
-            DisableAppPinIfDevicePinIsSet           = $policy.disableAppPinIfDevicePinIsSet
-            MaximumPinRetries                       = $policy.MaximumPinRetries
-            SimplePinBlocked                        = $policy.SimplePinBlocked
-            MinimumPinLength                        = $policy.MinimumPinLength
-            PinCharacterSet                         = $policy.PinCharacterSet
-            AllowedDataStorageLocations             = $policy.AllowedDataStorageLocations
-            ContactSyncBlocked                      = $policy.ContactSyncBlocked
-            PeriodBeforePinReset                    = $policy.PeriodBeforePinReset
-            PrintBlocked                            = $policy.PrintBlocked
-            FingerprintBlocked                      = $policy.FingerprintBlocked
-            Assignments                             = $assignmentsArray
-            ExcludedGroups                          = $exclusionArray
-            Apps                                    = $appsArray
-            Ensure                                  = "Present"
-            Credential                              = $Credential
-            ApplicationId                           = $ApplicationId
-            ApplicationSecret                       = $ApplicationSecret
-            TenantId                                = $TenantId
-            CertificateThumbprint                   = $CertificateThumbprint
+
+        $policy = @{}
+
+        $credentialParams = @(
+            'Credential',
+            'ApplicationId',
+            'TenantId',
+            'ApplicationSecret',
+            'CertificateThumbprint'
+        )
+
+        $credentialParams | foreach { $policy.add($_, (get-variable $_).value) }
+
+        $ComplexParameters = @{
+            'Apps' = $appsArray;
+            'Assignments' = $assignmentsArray;
+            'ExcludedGroups' = $exclusionArray
         }
+
+        $AllParams = (get-command Get-TargetResource).Parameters
+
+        foreach ($property in ($policyInfo | get-member -MemberType properties))
+        {
+            if (!($ComplexParameters.keys -contains $property.name) -and ($AllParams.keys -contains $property.name))
+            {
+                $policy.add($property.name, $policyInfo.($property.name) )
+            }
+        }
+
+        $ComplexParameters.keys | foreach { $policy.add($_, $ComplexParameters.$_ ) }
+        $policy.add('Ensure', 'Present' )
+        $Policy.add('id', $policyInfo.id)
+
+        return $policy
     }
     catch
     {
+        write-Verbose -message "ERROR on get-targetresource for $displayName"
         try
         {
             Write-Verbose -Message $_
@@ -342,30 +313,6 @@ function Set-TargetResource
 
         [Parameter()]
         [System.Boolean]
-        $IsAssigned,
-
-        [Parameter()]
-        [System.String]
-        $ManagedBrowser,
-
-        [Parameter()]
-        [System.String]
-        $MinimumRequiredAppVersion,
-
-        [Parameter()]
-        [System.String]
-        $MinimumWarningAppVersion,
-
-        [Parameter()]
-        [System.String]
-        $MinimumRequiredOSVersion,
-
-        [Parameter()]
-        [System.String]
-        $MinimumWarningOSVersion,
-
-        [Parameter()]
-        [System.Boolean]
         $ManagedBrowserToOpenLinksRequired,
 
         [Parameter()]
@@ -421,15 +368,20 @@ function Set-TargetResource
         $FingerprintBlocked,
 
         [Parameter()]
-        [System.String[]]
+        [ValidateSet('allApps', 'allMicrosoftApps', 'allCoreMicrosoftApps', 'selectedPublicApps' )]
+        [System.String]
+        $AppGroupType,
+
+        [Parameter()]
+        [System.Array[]]
         $Apps,
 
         [Parameter()]
-        [System.String[]]
+        [System.Array[]]
         $Assignments,
 
         [Parameter()]
-        [System.String[]]
+        [System.Array[]]
         $ExcludedGroups,
 
         [Parameter(Mandatory = $true)]
@@ -476,43 +428,133 @@ function Set-TargetResource
 
     if ($currentPolicy.Ensure -eq "ERROR")
     {
-
         Throw 'Error when searching for current policy details - Please check verbose output for further detail'
-
     }
 
-    $setParams = $PSBoundParameters
-    $setParams.Remove("Ensure") | Out-Null
-    $setParams.Remove("Credential") | Out-Null
+    $setParams = @{}
+    $PSBoundParameters.Remove("Ensure") | Out-Null
+    $PSBoundParameters.Remove("Credential") | Out-Null
+    $PSBoundParameters.Remove("ApplicationID") | Out-Null
+    $PSBoundParameters.Remove("ApplicationSecret") | Out-Null
+    $PSBoundParameters.Remove("TenantID") | Out-Null
+    $PSBoundParameters.Remove("CertificateThumbprint") | Out-Null
+
+    $AppsHash = set-AppsHash -AppGroupType $AppGroupType -apps $apps
+
+    $assignmentsArray = @()
+    $appsarray = @()
+    $configstring = "`r`nConfiguration To Be Applied:`r`n"
+
+    $ComplexParameters = @(
+        'AppGroupType',
+        'Apps',
+        'Assignments',
+        'ExcludedGroups'
+    )
+
     if ($Ensure -eq 'Present' -and $currentPolicy.Ensure -eq 'Absent')
     {
+        $graphparams = get-inputParamsList -commandName 'New-MgDeviceAppMgtAndroidManagedAppProtection'
         Write-Verbose -Message "Creating new Android App Protection Policy {$DisplayName}"
-        $JsonContent = Get-M365DSCIntuneAppProtectionPolicyAndroidJSON -Parameters $PSBoundParameters
-        Write-Verbose -Message "JSON: $JsonContent"
-        New-M365DSCIntuneAppProtectionPolicyAndroid -JSONContent $JsonContent
-
-        $policyInfo = Get-MgDeviceAppManagementAndroidManagedAppProtection -Filter "displayName eq '$DisplayName'" `
-            -ErrorAction Stop
-        $assignmentJSON = Get-M365DSCIntuneAppProtectionPolicyAndroidAssignmentJson -Assignments $Assignments `
-            -Exclusions $ExcludedGroups
-
-        Set-M365DSCIntuneAppProtectionPolicyAndroidAssignment -JsonContent $assignmentJSON `
-            -PolicyId $policyInfo.Id
     }
     elseif ($Ensure -eq 'Present' -and $currentPolicy.Ensure -eq 'Present')
     {
         Write-Verbose -Message "Updating existing Android App Protection Policy {$DisplayName}"
-        $policyInfo = Get-MgDeviceAppManagementAndroidManagedAppProtection -Filter "displayName eq '$DisplayName'" `
-            -ErrorAction Stop
+        $setParams.add('AndroidManagedAppProtectionId', (Get-MgDeviceAppManagementAndroidManagedAppProtection -Filter "displayName eq '$DisplayName'").id)
+        $graphparams = get-inputParamsList -commandName 'Update-MgDeviceAppMgtAndroidManagedAppProtection'
+    }
+    elseif ($Ensure -eq 'Absent' -and $currentPolicy.Ensure -eq 'Present'){}
 
-        $JsonContent = Get-M365DSCIntuneAppProtectionPolicyAndroidJSON -Parameters $PSBoundParameters `
-            -IncludeApps $false
-        Set-M365DSCIntuneAppProtectionPolicyAndroid -JSONContent $JsonContent `
-            -PolicyId ($policyInfo.id)
+    # construct command
+    write-verbose -message "Preparing input parameters..."
+    foreach ($param in $PSBoundParameters.keys)
+    {
+        if ( $graphparams.containskey($param) -and !($ComplexParameters -contains $param) )
+        {
+            switch ($graphparams.$param.ParameterType.name)
+            {
+                'TimeSpan'
+                {
+                    $setParams.add($param, (set-TimeSpan -duration $PSBoundParameters.$param))
+                    $configstring += ($param + ':' + ($setParams.$param.tostring()) + "`r`n")
+                }
 
-        $appJSON = Get-M365DSCIntuneAppProtectionPolicyAndroidAppsJSON -Parameters $PSBoundParameters
-        Set-M365DSCIntuneAppProtectionPolicyAndroidApps -JSONContent $appJSON `
-            -PolicyId $policyInfo.Id
+                'string'
+                {
+                    $setParams.add($param, $PSBoundParameters.$param)
+                    $configstring += ($param + ':' + $setParams.$param + "`r`n")
+                }
+
+                default
+                {
+                    $setParams.add($param, $PSBoundParameters.$param)
+                    $configstring += ($param + ':' + $setParams.$param + "`r`n")
+                }
+            }
+        }
+
+        else
+        {
+            #write-verbose -message "Cannot specify this in graph cmdlet: $param"
+        }
+    }
+
+    foreach ($param in $ComplexParameters)
+    {
+        switch ($param)
+        {
+            'AppGroupType'
+            {
+                $configstring += ('AppGroupType:' + $appshash.AppGroupType + "`r`n")
+            }
+
+            'Apps'
+            {
+                if ($appshash.AppGroupType -eq 'selectedPublicApps' )
+                {
+                    $appshash.Apps | foreach {
+                        if ($_ -ne $null) {$appsarray+= set-JSONstring -id $_ -type $param}
+                    }
+                    $configstring += ($param + ":`r`n" +($appshash.Apps | out-string) + "`r`n" )
+                }
+            }
+
+            'Assignments'
+            {
+                $PSBoundParameters.$param | foreach {
+                    if ($_ -ne $null) {$assignmentsArray+= set-JSONstring -id $_ -type $param}
+                }
+                $configstring += ( $param + ":`r`n" +($PSBoundParameters.$param | out-string) + "`r`n" )
+            }
+
+            'ExcludedGroups'
+            {
+                $PSBoundParameters.$param | foreach {
+                    if ($_ -ne $null) {$assignmentsArray+= set-JSONstring -id $_ -type $param}
+                }
+                $configstring += ($param + ":`r`n" +($PSBoundParameters.$param | out-string) + "`r`n" )
+            }
+        }
+    }
+
+    write-verbose -message $configstring
+
+    if ($Ensure -eq 'Present' -and $currentPolicy.Ensure -eq 'Absent')
+    {
+        Write-Verbose -Message 'Setting up policy via graph...'
+        $setParams.add('Assignments', $assignmentsArray)
+        $newpolicy = New-MgDeviceAppMgtAndroidManagedAppProtection @setParams
+        $setParams.add('AndroidManagedAppProtectionId', $newpolicy.Id)
+
+    }
+
+    elseif ($Ensure -eq 'Present' -and $currentPolicy.Ensure -eq 'Present')
+    {
+        Write-Verbose -Message 'Amending policy via graph...'
+        Update-MgDeviceAppMgtAndroidManagedAppProtection @setParams
+
+        Write-Verbose -Message 'Setting Group Assignments...'
+        set-MgDeviceAppMgtTargetedManagedAppConfiguration -TargetedManagedAppConfigurationId $setParams.AndroidManagedAppProtectionId -Assignments $assignmentsArray
 
     }
     elseif ($Ensure -eq 'Absent' -and $currentPolicy.Ensure -eq 'Present')
@@ -522,6 +564,13 @@ function Set-TargetResource
             -ErrorAction Stop
         Remove-MgDeviceAppManagementAndroidManagedAppProtection -AndroidManagedAppProtectionId $policyInfo.id
     }
+
+    if ( $ensure -ne 'Absent')
+    {
+        write-verbose -message ("Setting Application values of type: " + $AppsHash.AppGroupType)
+        Invoke-MgTargetDeviceAppMgtTargetedManagedAppConfigurationApp -TargetedManagedAppConfigurationId $setParams.AndroidManagedAppProtectionId -Apps $appsarray -AppGroupType $AppsHash.AppGroupType
+    }
+
 }
 
 function Test-TargetResource
@@ -572,30 +621,6 @@ function Test-TargetResource
 
         [Parameter()]
         [System.Boolean]
-        $IsAssigned,
-
-        [Parameter()]
-        [System.String]
-        $ManagedBrowser,
-
-        [Parameter()]
-        [System.String]
-        $MinimumRequiredAppVersion,
-
-        [Parameter()]
-        [System.String]
-        $MinimumWarningAppVersion,
-
-        [Parameter()]
-        [System.String]
-        $MinimumRequiredOSVersion,
-
-        [Parameter()]
-        [System.String]
-        $MinimumWarningOSVersion,
-
-        [Parameter()]
-        [System.Boolean]
         $ManagedBrowserToOpenLinksRequired,
 
         [Parameter()]
@@ -649,6 +674,11 @@ function Test-TargetResource
         [Parameter()]
         [System.Boolean]
         $FingerprintBlocked,
+
+        [Parameter()]
+        [ValidateSet('allApps', 'allMicrosoftApps', 'allCoreMicrosoftApps', 'selectedPublicApps' )]
+        [System.String]
+        $AppGroupType,
 
         [Parameter()]
         [System.String[]]
@@ -702,6 +732,8 @@ function Test-TargetResource
 
     $CurrentValues = Get-TargetResource @PSBoundParameters
 
+    if ($CurrentValues.AppGroupType -ne 'SelectedPublicApps') { $CurrentValues.Apps = @() }
+
     if ($CurrentValues.Ensure -eq "ERROR")
     {
 
@@ -709,19 +741,68 @@ function Test-TargetResource
 
     }
 
-    Write-Verbose -Message "Current Values: $(Convert-M365DscHashtableToString -Hashtable $CurrentValues)"
-    Write-Verbose -Message "Target Values: $(Convert-M365DscHashtableToString -Hashtable $PSBoundParameters)"
+    $ValuesToCheck = @{}
 
-    $ValuesToCheck = $PSBoundParameters
-    $ValuesToCheck.Remove('Credential') | Out-Null
-    $ValuesToCheck.Remove('ApplicationId') | Out-Null
-    $ValuesToCheck.Remove('TenantId') | Out-Null
-    $ValuesToCheck.Remove('ApplicationSecret') | Out-Null
+    $ComplexParameters = @(
+        'AppGroupType',
+        'Apps',
+        'Assignments',
+        'ExcludedGroups'
+    )
+
+    $credentialParams = @(
+        'Credential',
+        'ApplicationId',
+        'TenantId',
+        'ApplicationSecret',
+        'CertificateThumbprint'
+    )
+
+    ($allparams = (get-command Test-TargetResource).Parameters).keys | foreach {
+
+        if(!($PSBoundParameters.keys -contains $allparams.$_.name) -and ($allparams.$_.Aliases.count -eq 0) -and !($credentialParams -contains $_) )
+        {
+            if ($ComplexParameters -contains $allparams.$_.name)
+            {
+                write-host 'Unspecified Parameter in Config:' $allparams.$_.name  "- Current Value is:" $CurrentValues.$_ `
+                "`r`nNOTE:" $allparams.$_.name "interacts with other values - not specifying may lead to unexpected output"
+                # set to '' at this stage - the apps values are configured below
+                $ValuesToCheck.add( ($allparams.$_.name) , '')
+            }
+            else
+            {
+                write-host 'Unspecified Parameter in Config:' $allparams.$_.name  "Current Value Will be retained:" $CurrentValues.$_
+            }
+        }
+    }
+
+    $AppsHash = set-AppsHash -AppGroupType $AppGroupType -apps $apps
+    $PSBoundParameters.AppGroupType = $AppsHash.AppGroupType
+    $PSBoundParameters.Apps = $AppsHash.Apps
+
+    $credentialParams | foreach {
+        $CurrentValues.Remove($_) | out-null
+        $PSBoundParameters.Remove($_) | out-null
+    }
+
+    $PSBoundParameters.keys | foreach {
+        if ($currentvalues.$_ -ne $null)
+        {
+            if ($currentvalues.$_.gettype().name -eq 'TimeSpan') { $ValuesToCheck.$_ = set-timespan -duration $PSBoundParameters.$_ }
+            else{ $ValuesToCheck.$_ = $PSBoundParameters.$_ }
+        }
+        else{ $ValuesToCheck.$_ = $PSBoundParameters.$_ }
+    }
+
+
+    Write-Verbose -Message "Current Values: $((Convert-M365DscHashtableToString -Hashtable $CurrentValues) -replace ';', "`r`n")"
+    Write-Verbose -Message "Target Values: $((Convert-M365DscHashtableToString -Hashtable $ValuesToCheck) -replace ';', "`r`n")"
 
     $TestResult = Test-M365DSCParameterState -CurrentValues $CurrentValues `
         -Source $($MyInvocation.MyCommand.Source) `
-        -DesiredValues $PSBoundParameters `
-        -ValuesToCheck $ValuesToCheck.Keys
+        -DesiredValues $ValuesToCheck `
+        -ValuesToCheck $ValuesToCheck.Keys `
+        -verbose
 
     Write-Verbose -Message "Test-TargetResource returned $TestResult"
 
@@ -754,7 +835,7 @@ function Export-TargetResource
         [System.String]
         $CertificateThumbprint
     )
-    $ConnectionMode = New-M365DSCConnection -Workload 'MicrosoftGraph' `
+    $ConnectionMode = New-M365DSCConnection -Workload 'MicrosoftGraph' -ProfileName v1.0 `
         -InboundParameters $PSBoundParameters
 
     #Ensure the proper dependencies are installed in the current environment.
@@ -795,6 +876,7 @@ function Export-TargetResource
                 CertificateThumbprint = $CertificateThumbprint
             }
             $Results = Get-TargetResource @Params
+            $Results.remove('id') | out-null
             $Results = Update-M365DSCExportAuthenticationResults -ConnectionMode $ConnectionMode `
                 -Results $Results
             $currentDSCBlock = Get-M365DSCExportContentForResource -ResourceName $ResourceName `
@@ -837,319 +919,124 @@ function Export-TargetResource
     }
 }
 
-function Get-M365DSCIntuneAppProtectionPolicyAndroid
+function get-inputParamsList
 {
-    [CmdletBinding()]
-    [OutputType([PSCustomObject])]
     param(
-        [Parameter(Mandatory = $true)]
-        [System.String]
-        $PolicyId
+        [string]$commandName
+    )
+
+    $Graphparams = (get-command $commandName).Parameters
+
+    return $Graphparams
+
+}
+
+function set-Timespan
+{
+    param(
+        [string]$duration
     )
     try
     {
-        $Url = "https://graph.microsoft.com/beta/deviceAppManagement/AndroidManagedAppProtections('$PolicyId')/`?expand=apps,assignments"
-        $response = Invoke-MgGraphRequest -Method Get `
-            -Uri $Url
-        return $response
+        if ($duration.startswith('P'))
+        {
+            $timespan = [System.Xml.XmlConvert]::ToTimeSpan($duration)
+        }
+        else
+        {
+            $timespan = [TimeSpan]$duration
+        }
     }
     catch
     {
-        Write-Verbose -Message $_
-        $tenantIdValue = $Credential.UserName.Split('@')[1]
-        Add-M365DSCEvent -Message $_ -EntryType 'Error' `
-            -EventID 1 -Source $($MyInvocation.MyCommand.Source) `
-            -TenantId $tenantIdValue
+        throw "Problem converting input to a timespan - If configuration document is using iso8601 string (e.g. PT15M) try using new-timespan (e.g. new-timespan -minutes 15)"
     }
-    return $null
+    return $timespan
 }
 
-function Get-M365DSCIntuneAppProtectionPolicyAndroidJSON
+function set-JSONstring
 {
-    [CmdletBinding()]
-    [OutputType([System.String])]
     param(
-        [Parameter(Mandatory = $true)]
-        [System.Collections.Hashtable]
-        $Parameters,
-
-        [Parameter()]
-        [System.Boolean]
-        $IncludeApps = $true
+        [string]$id,
+        [string]$type
     )
 
-    #region AllowedDataStorageLocations
-    $allowedDataStorageLocations = "["
-    $foundOne = $false
-    foreach ($allowedLocation in $Parameters.AllowedDataStorageLocations)
-    {
-        $foundOne = $true
-        $allowedDataStorageLocations += "`r`n`"$allowedLocation`","
-    }
-    if ($foundOne)
-    {
-        $allowedDataStorageLocations = $allowedDataStorageLocations.TrimEnd(',') + " `r`n"
-    }
-    $allowedDataStorageLocations += "],"
-    #endregion
+    $JsonContent = ''
 
-    #region Apps
-    $appsValue = "["
-    $foundOne = $false
-    foreach ($app in $Parameters.Apps)
+    switch ($type)
     {
-        $foundOne = $true
 
-        $appsValue += @"
-            `r`n{
-                "id":"$($app)",
-                "mobileAppIdentifier": {
-                    "@odata.type": "#microsoft.graph.androidMobileAppIdentifier",
-                    "packageId": "$app"
+        'Apps'
+        {
+            $JsonContent = @"
+                    {
+                    "id":"$($id)",
+                    "mobileAppIdentifier": {
+                        "@odata.type": "#microsoft.graph.AndroidMobileAppIdentifier",
+                        "packageId": "$id"
+                    }
                 }
-            },
-"@
-    }
-    if ($foundOne)
-    {
-        $appsValue = $appsValue.TrimEnd(',') + " `r`n"
-    }
-    $appsValue += "]"
-    #endregion
-    $JsonContent = @"
-    {
-        "@odata.type": "#microsoft.graph.androidManagedAppProtection",
-        "displayName": "$($Parameters.DisplayName)",
-        "description": "$($Parameters.Description)",
-        "periodOfflineBeforeAccessCheck": "$($Parameters.PeriodOfflineBeforeAccessCheck)",
-        "periodOnlineBeforeAccessCheck": "$($Parameters.PeriodOnlineBeforeAccessCheck)",
-        "allowedInboundDataTransferSources": "$($Parameters.AllowedInboundDataTransferSources)",
-        "allowedOutboundDataTransferDestinations": "$($Parameters.AllowedOutboundDataTransferDestinations)",
-        "organizationalCredentialsRequired": $($Parameters.OrganizationalCredentialsRequired.ToString().ToLower()),
-        "allowedOutboundClipboardSharingLevel": "$($Parameters.AllowedOutboundClipboardSharingLevel)",
-        "dataBackupBlocked": $($Parameters.DataBackupBlocked.ToString().ToLower()),
-        "deviceComplianceRequired": $($Parameters.DeviceComplianceRequired.ToString().ToLower()),
-        "managedBrowserToOpenLinksRequired": $($Parameters.ManagedBrowserToOpenLinksRequired.ToString().ToLower()),
-        "saveAsBlocked": $($Parameters.SaveAsBlocked.ToString().ToLower()),
-        "periodOfflineBeforeWipeIsEnforced": "$($Parameters.PeriodOfflineBeforeWipeIsEnforced)",
-        "pinRequired": $($Parameters.PinRequired.ToString().ToLower()),
-        "disableAppPinIfDevicePinIsSet": $($Parameters.DisableAppPinIfDevicePinIsSet.ToString().ToLower()),
-        "maximumPinRetries": $($Parameters.MaximumPinRetries),
-        "simplePinBlocked": $($Parameters.SimplePinBlocked.ToString().ToLower()),
-        "minimumPinLength": $($Parameters.MinimumPinLength),
-        "pinCharacterSet": "$($Parameters.PinCharacterSet)",
-        "contactSyncBlocked": $($Parameters.ContactSyncBlocked.ToString().ToLower()),
-        "periodBeforePinReset": "$($Parameters.PeriodBeforePinReset)",
-        "printBlocked": $($Parameters.PrintBlocked.ToString().ToLower()),
-        "fingerprintBlocked": $($Parameters.FingerprintBlocked.ToString().ToLower()),
-        "allowedDataStorageLocations": $allowedDataStorageLocations
 "@
 
-    if ($IncludeApps)
-    {
-        $JSOnContent += "`"apps`":$appsValue`r`n"
-    }
-    $JsonContent += "}"
-    return $JsonContent
-}
-
-function Get-M365DSCIntuneAppProtectionPolicyAndroidAppsJSON
-{
-    [CmdletBinding()]
-    [OutputType([System.String])]
-    param(
-        [Parameter(Mandatory = $true)]
-        [System.Collections.Hashtable]
-        $Parameters
-    )
-
-    #region Apps
-    $appsValue = "["
-    $foundOne = $false
-    foreach ($app in $Parameters.Apps)
-    {
-        $foundOne = $true
-
-        $appsValue += @"
-            `r`n{
-                "id":"$($app)",
-                "mobileAppIdentifier": {
-                    "@odata.type": "#microsoft.graph.AndroidMobileAppIdentifier",
-                    "packageId": "$app"
-                }
-            },
+        }
+        'Assignments'
+        {
+            $JsonContent = @"
+            {
+            "id":"$($id)_incl",
+            "target": {
+                        "@odata.type": "#microsoft.graph.groupAssignmentTarget",
+                        "groupId": "$($id)"
+                    }
+            }
 "@
-    }
-    if ($foundOne)
-    {
-        $appsValue = $appsValue.TrimEnd(',') + " `r`n"
-    }
-    $appsValue += "]"
-    #endregion
-
-    $JsonContent = @"
-    {
-        "apps": $appsValue
-    }
+        }
+        'ExcludedGroups'
+        {
+            $JsonContent = @"
+            {
+            "id":"$($id)_excl",
+            "target": {
+                        "@odata.type": "#microsoft.graph.exclusionGroupAssignmentTarget",
+                        "groupId": "$($id)"
+                    }
+            }
 "@
-    return $JsonContent
-}
 
-function Get-M365DSCIntuneAppProtectionPolicyAndroidAssignmentJSON
-{
-    [CmdletBinding()]
-    [OutputType([System.String])]
-    param(
-        [Parameter(Mandatory = $true)]
-        [System.String[]]
-        $Assignments,
-
-        [Parameter(Mandatory = $false)]
-        [System.String[]]
-        $Exclusions
-    )
-
-    $JsonContent = "{`r`n"
-    $JsonContent += "`"assignments`":[`r`n"
-    foreach ($assignment in $Assignments)
-    {
-        $JsonContent += "    {`"target`":{`r`n"
-        $JsonContent += "        `"groupId`":`"$assignment`",`r`n"
-        $JsonContent += "        `"@odata.type`":`"#microsoft.graph.groupAssignmentTarget`"`r`n"
-        $JsonContent += "    }},"
+        }
     }
-    foreach ($exclusion in $Exclusions)
-    {
-        $JsonContent += "    {`"target`":{`r`n"
-        $JsonContent += "        `"groupId`":`"$exclusion`",`r`n"
-        $JsonContent += "        `"@odata.type`":`"#microsoft.graph.exclusionGroupAssignmentTarget`"`r`n"
-        $JsonContent += "    }},"
-    }
-    $JsonContent = $JsonContent.TrimEnd(',')
-    $JsonContent += "]`r`n"
-    $JsonContent += "`r`n}"
 
     return $JsonContent
+
 }
 
-function New-M365DSCIntuneAppProtectionPolicyAndroid
+function set-AppsHash
 {
-    [CmdletBinding()]
-    param(
-        [Parameter(Mandatory = $true)]
-        [System.String]
-        $JSONContent
+    param (
+        [string]$AppGroupType,
+        [array]$Apps
     )
-    try
+
+   if ($AppGroupType -eq '')
     {
-        $Url = 'https://graph.microsoft.com/beta/deviceAppManagement/managedAppPolicies'
-        Write-Verbose -Message "Creating new Android App Protection policy with JSON payload: `r`n$JSONContent"
-        Invoke-MgGraphRequest -Method POST `
-            -Uri $Url `
-            -Body $JSONContent `
-            -Headers @{"Content-Type" = "application/json" } | Out-Null
+        if ($apps.count -eq 0 ) { $AppGroupType = 'allApps' }
+        else { $AppGroupType = 'selectedPublicApps' }
+        write-verbose -message "setting AppGroupType to $AppGroupType"
     }
-    catch
+
+    $appsarray = @()
+    if ($AppGroupType -eq 'selectedPublicApps' )
     {
-        Write-Verbose -Message $_
-        $tenantIdValue = $Credential.UserName.Split('@')[1]
-        Add-M365DSCEvent -Message $_ -EntryType 'Error' `
-            -EventID 1 -Source $($MyInvocation.MyCommand.Source) `
-            -TenantId $tenantIdValue
+        $appsarray = $apps
     }
+
+    $AppsHash = @{
+        'AppGroupType' = $AppGroupType;
+        'Apps' = $appsarray
+    }
+
+    return $AppsHash
 }
 
-function Set-M365DSCIntuneAppProtectionPolicyAndroid
-{
-    [CmdletBinding()]
-    param(
-        [Parameter(Mandatory = $true)]
-        [System.String]
-        $JSONContent,
-
-        [Parameter(Mandatory = $true)]
-        [System.String]
-        $PolicyId
-    )
-    try
-    {
-        $Url = "https://graph.microsoft.com/beta/deviceAppManagement/androidManagedAppProtections('$PolicyId')/"
-        Write-Verbose -Message "Updating Android App Protection policy with JSON payload: `r`n$JSONContent"
-        Invoke-MgGraphRequest -Method PATCH `
-            -Uri $Url `
-            -Body $JSONContent `
-            -Headers @{"Content-Type" = "application/json" } | Out-Null
-    }
-    catch
-    {
-        Write-Verbose -Message $_
-        $tenantIdValue = $Credential.UserName.Split('@')[1]
-        Add-M365DSCEvent -Message $_ -EntryType 'Error' `
-            -EventID 1 -Source $($MyInvocation.MyCommand.Source) `
-            -TenantId $tenantIdValue
-    }
-}
-
-function Set-M365DSCIntuneAppProtectionPolicyAndroidApps
-{
-    [CmdletBinding()]
-    param(
-        [Parameter(Mandatory = $true)]
-        [System.String]
-        $JSONContent,
-
-        [Parameter(Mandatory = $true)]
-        [System.String]
-        $PolicyId
-    )
-    try
-    {
-        $Url = "https://graph.microsoft.com/beta/deviceAppManagement/managedAppPolicies/$PolicyId/targetApps"
-        Write-Verbose -Message "Updating Apps for Android App Protection policy with JSON payload: `r`n$JSONContent"
-        Invoke-MgGraphRequest -Method POST `
-            -Uri $Url `
-            -Body $JSONContent `
-            -Headers @{"Content-Type" = "application/json" } | Out-Null
-    }
-    catch
-    {
-        Write-Verbose -Message $_
-        $tenantIdValue = $Credential.UserName.Split('@')[1]
-        Add-M365DSCEvent -Message $_ -EntryType 'Error' `
-            -EventID 1 -Source $($MyInvocation.MyCommand.Source) `
-            -TenantId $tenantIdValue
-    }
-}
-
-function Set-M365DSCIntuneAppProtectionPolicyAndroidAssignment
-{
-    [CmdletBinding()]
-    param(
-        [Parameter(Mandatory = $true)]
-        [System.String]
-        $JSONContent,
-
-        [Parameter(Mandatory = $true)]
-        [System.String]
-        $PolicyId
-    )
-    try
-    {
-        $Url = "https://graph.microsoft.com/beta/deviceAppManagement/androidManagedAppProtections('$PolicyId')/assign"
-        Write-Verbose -Message "Group Assignment for Android App Protection policy with JSON payload: `r`n$JSONContent"
-        Invoke-MgGraphRequest -Method POST `
-            -Uri $Url `
-            -Body $JSONContent `
-            -Headers @{"Content-Type" = "application/json" } | Out-Null
-    }
-    catch
-    {
-        Write-Verbose -Message $_
-        $tenantIdValue = $Credential.UserName.Split('@')[1]
-        Add-M365DSCEvent -Message $_ -EntryType 'Error' `
-            -EventID 1 -Source $($MyInvocation.MyCommand.Source) `
-            -TenantId $tenantIdValue
-    }
-}
 
 Export-ModuleMember -Function *-TargetResource